Stop add learner submit event propogation (611)
Pressing enter while the email form control is focused causes
the page to refresh and the new learner not to be added because
the submit event causes the form to be submitted and there is
nothing to stop propagation of the event.

Fix by moving the update logic from a button click handler to a
form submit handler and stopping propagation of the event.